updated exercise 2 file structure and tests
