+++ src/gql/gql.ts
@@ -13,11 +13,15 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
  * Therefore it is highly recommended to use the babel-plugin for production.
  */
 const documents = {
+    "\n    query GetState($name: String) {\n        states(name: $name) {\n            id\n            name\n            key\n            slug\n        }\n    }\n": types.GetStateDocument,
     "\n    query GetTrade($name: String) {\n        states(name: $name) {\n            tradeSummary {\n                name\n                totalDollarAmount\n                totalTons\n                statesByDollars {\n                    name\n                    amount\n                }\n                statesByTons {\n                    name\n                    amount\n                }\n            }\n        }\n    }\n": types.GetTradeDocument,
     "\n    query Economy(\n        $name: String\n        $production: Boolean!\n        $employment: Boolean!\n        $trade: Boolean!\n    ) {\n        states(name: $name) {\n            name\n            productionSummary @include(if: $production) {\n                productionTypeByDollars {\n                    name\n                    amount\n                }\n                productionTypeByTons {\n                    name\n                    amount\n                }\n            }\n            employmentSummary @include(if: $employment) {\n                topIndustryByAverageSalary {\n                    industry\n                    averageSalary\n                }\n            }\n            tradeSummary @include(if: $trade) {\n                totalDollarAmount\n                totalTons\n                statesByDollars {\n                    name\n                    amount\n                }\n                statesByTons {\n                    name\n                    amount\n                }\n            }\n        }\n    }\n": types.EconomyDocument,
-    "\n    query GetState($name: String) {\n        states(name: $name) {\n            id\n            name\n            key\n            slug\n        }\n    }\n": types.GetStateDocument,
 };
 
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n    query GetState($name: String) {\n        states(name: $name) {\n            id\n            name\n            key\n            slug\n        }\n    }\n"): (typeof documents)["\n    query GetState($name: String) {\n        states(name: $name) {\n            id\n            name\n            key\n            slug\n        }\n    }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -26,10 +30,6 @@ export function graphql(source: "\n    query GetTrade($name: String) {\n        
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n    query Economy(\n        $name: String\n        $production: Boolean!\n        $employment: Boolean!\n        $trade: Boolean!\n    ) {\n        states(name: $name) {\n            name\n            productionSummary @include(if: $production) {\n                productionTypeByDollars {\n                    name\n                    amount\n                }\n                productionTypeByTons {\n                    name\n                    amount\n                }\n            }\n            employmentSummary @include(if: $employment) {\n                topIndustryByAverageSalary {\n                    industry\n                    averageSalary\n                }\n            }\n            tradeSummary @include(if: $trade) {\n                totalDollarAmount\n                totalTons\n                statesByDollars {\n                    name\n                    amount\n                }\n                statesByTons {\n                    name\n                    amount\n                }\n            }\n        }\n    }\n"): (typeof documents)["\n    query Economy(\n        $name: String\n        $production: Boolean!\n        $employment: Boolean!\n        $trade: Boolean!\n    ) {\n        states(name: $name) {\n            name\n            productionSummary @include(if: $production) {\n                productionTypeByDollars {\n                    name\n                    amount\n                }\n                productionTypeByTons {\n                    name\n                    amount\n                }\n            }\n            employmentSummary @include(if: $employment) {\n                topIndustryByAverageSalary {\n                    industry\n                    averageSalary\n                }\n            }\n            tradeSummary @include(if: $trade) {\n                totalDollarAmount\n                totalTons\n                statesByDollars {\n                    name\n                    amount\n                }\n                statesByTons {\n                    name\n                    amount\n                }\n            }\n        }\n    }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n    query GetState($name: String) {\n        states(name: $name) {\n            id\n            name\n            key\n            slug\n        }\n    }\n"): (typeof documents)["\n    query GetState($name: String) {\n        states(name: $name) {\n            id\n            name\n            key\n            slug\n        }\n    }\n"];
 
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
